requirements.txt
openai
python-dotenv
pandas
openpyxl
tiktoken